Add method to get adjacent hex
This helper method is the beginning of the implementation of
building roads, settlements, and cities. A lot of helper functions
are going to be needed since each of these tasks are very involved.
Another change that was made was to create a new HexPoint data
structure to keep track of the hex points without using the Point
class since that uses x and y coordinates, which doesn't make that
much sense for this project.